third_party/blink/web_tests/fast/events/event-fire-order-expected.txt

This page tests event listener fire order for a few objects that had it wrong in the past.

If the test passes, you'll see a series of PASS messages below.

PASS: result should be f1,f2 and is.
PASS: result should be f1,f2 and is.
PASS: result should be f1,f2 and is.
